The first step - transferring Dto models to Services
