+++ OstreCWEB/OstreCWEB.Services/Mapping/StoryBuilderProfile.cs
@@ -1,0 +1,8 @@
+﻿using AutoMapper;
+
+namespace OstreCWEB.Services.Mapping
+{
+    internal class StoryBuilderProfile : Profile
+    {
+    }
+}
