+++ OstreCWEB/OstreCWEB.Services/StoryServices/ModelsView/ChoiceCreatorView.cs
@@ -1,4 +1,4 @@
-﻿namespace OstreCWEB.ViewModel.StoryBuilder
+﻿namespace OstreCWEB.Services.StoryServices.ModelsView
 {
     public class ChoiceCreatorView
     {
