+++ OstreCWEB/OstreCWEB.Services/StoryServices/ModelsView/EditParagraphView.cs
@@ -2,7 +2,7 @@
 using OstreCWEB.DomainModels.StoryModels.Enums;
 using System.ComponentModel.DataAnnotations;
 
-namespace OstreCWEB.ViewModel.StoryBuilder
+namespace OstreCWEB.Services.StoryServices.ModelsView
 {
     public class EditParagraphView
     {
@@ -13,17 +13,19 @@ namespace OstreCWEB.ViewModel.StoryBuilder
         public bool RestoreRest { get; set; }
 
         // Paragraph Fight properties
-        public List<EnemyInParagraphView> ParagraphEnemies { get; set; }
+        public List<OstreCWEB.Services.StoryServices.ModelsView.EnemyInParagraphView> ParagraphEnemies { get; set; }
 
         // ParagraphTest Properties
         [Display(Name = "Ability Score")]
         public AbilityScores AbilityScores { get; set; }
+
         [Display(Name = "Test Difficulty")]
         public TestDifficulty TestDifficulty { get; set; }
 
         // Db relations properties
         //public List<ParagraphItem> ParagraphItems { get; set; }
         public int FightPropId { get; set; }
+
         public int StoryId { get; set; }
     }
 }
