+++ OstreCWEB/OstreCWEB.Services/StoryServices/ModelsView/ParagraphDetailsView.cs
@@ -1,6 +1,6 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace OstreCWEB.ViewModel.StoryBuilder
+namespace OstreCWEB.Services.StoryServices.ModelsView
 {
     public class ParagraphDetailsView
     {
