+++ OstreCWEB/OstreCWEB.Services/StoryServices/ModelsView/StoriesView.cs
@@ -1,6 +1,6 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace OstreCWEB.ViewModel.StoryBuilder
+namespace OstreCWEB.Services.StoryServices.ModelsView
 {
     public class StoriesView
     {
