+++ OstreCWEB/OstreCWEB.Services/StoryServices/ModelsView/StoryParagraphsView.cs
@@ -1,6 +1,6 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace OstreCWEB.ViewModel.StoryBuilder
+namespace OstreCWEB.Services.StoryServices.ModelsView
 {
     public class StoryParagraphsView
     {
@@ -10,6 +10,7 @@ namespace OstreCWEB.ViewModel.StoryBuilder
 
         [Display(Name = "Amount Of Paragraphs")]
         public int AmountOfParagraphs { get; set; }
+
         public List<ParagraphElementView> Paragraphs { get; set; }
 
         public int FirstParagraphId { get; set; }
