+++ OstreCWEB/OstreCWEB/ViewModel/Game/StartGameView.cs
@@ -1,25 +1,30 @@
 ﻿using OstreCWEB.ViewModel.Characters;
 using OstreCWEB.ViewModel.Identity;
-using OstreCWEB.ViewModel.StoryBuilder;
+using OstreCWEB.Services.StoryServices.ModelsView;
 using System.ComponentModel;
 
 namespace OstreCWEB.ViewModel.Game
 {
     public class StartGameView
     {
-        //This class will be merged with story reader next week. 
+        //This class will be merged with story reader next week.
         public UserView User { get; set; }
-        [DisplayName("Other characters")]
 
+        [DisplayName("Other characters")]
         public List<PlayableCharacterRow> OtherUsersCharacters { get; set; }
+
         [DisplayName("Other Stories")]
         public List<StoriesView> OtherUsersStories { get; set; }
+
         [DisplayName("Saved Game")]
         public UserParagraphView UserParagraph { get; set; }
+
         [DisplayName("Chosen Character")]
         public PlayableCharacterView ActiveCharacter { get; set; }
+
         [DisplayName("Chosen Story")]
         public StoriesView ActiveStory { get; set; }
+
         public StartGameView()
         {
             User = new UserView();
