+++ OstreCWEB/OstreCWEB/ViewModel/Identity/UserView.cs
@@ -1,5 +1,5 @@
 ﻿using OstreCWEB.ViewModel.Characters;
-using OstreCWEB.ViewModel.StoryBuilder;
+using OstreCWEB.Services.StoryServices.ModelsView;
 using System.ComponentModel;
 
 namespace OstreCWEB.ViewModel.Identity
@@ -8,12 +8,16 @@ namespace OstreCWEB.ViewModel.Identity
     {
         [DisplayName("Your Characters")]
         public List<PlayableCharacterRow> CharactersCreated { get; set; }
+
         [DisplayName("Your Stories")]
         public List<StoriesView> StoriesCreated { get; set; }
+
         [DisplayName("Your Saved Games")]
         public List<UserParagraphView> UserParagraphs { get; set; }
-        //  
+
+        //
         public string Name { get; set; }
-        public DateTime Created { get; set; } 
+
+        public DateTime Created { get; set; }
     }
 }
