+++ OstreCWEB/OstreCWEB/Views/Game/_StoryRow.cshtml
@@ -1,4 +1,4 @@
-﻿@model IEnumerable<OstreCWEB.ViewModel.StoryBuilder.StoriesView>
+﻿@model IEnumerable<OstreCWEB.Services.StoryServices.ModelsView.StoriesView>
 @{
     ViewData["Title"] = "_PlayableCharacterTable";
 }
