+++ OstreCWEB/OstreCWEB/Views/StoryBuilder/AddEnemyInParagraph.cshtml
@@ -1,4 +1,4 @@
-﻿@model OstreCWEB.ViewModel.StoryBuilder.EnemyInParagraphView
+﻿@model OstreCWEB.Services.StoryServices.ModelsView.EnemyInParagraphView
 
 @{
     ViewData["Title"] = "Add Enemy to Fight";
