+++ OstreCWEB/OstreCWEB/Views/StoryBuilder/ChoiceDetails.cshtml
@@ -1,5 +1,5 @@
 ﻿@using OstreCWEB.DomainModels.StoryModels.Enums
-@model OstreCWEB.ViewModel.StoryBuilder.ChoiceDetailsView
+@model OstreCWEB.Services.StoryServices.ModelsView.ChoiceDetailsView
 
 @{
     ViewData["Title"] = $"Story " + @Model.NameOfStory;
