+++ OstreCWEB/OstreCWEB/Views/StoryBuilder/ChooseSecondParagraph.cshtml
@@ -1,5 +1,5 @@
 ﻿@using OstreCWEB.DomainModels.StoryModels.Enums
-@model OstreCWEB.ViewModel.StoryBuilder.StoryParagraphsView
+@model OstreCWEB.Services.StoryServices.ModelsView.StoryParagraphsView
 
 @{
     ViewData["Title"] = $"Story " + @Model.Name;
