+++ OstreCWEB/OstreCWEB/Views/StoryBuilder/CreatChoice.cshtml
@@ -1,5 +1,5 @@
 ﻿@using OstreCWEB.DomainModels.StoryModels.Enums 
-@model OstreCWEB.ViewModel.StoryBuilder.ChoiceCreatorView
+@model OstreCWEB.Services.StoryServices.ModelsView.ChoiceCreatorView
 
 @{
     ViewData["Title"] = $"Choice Creator";
