+++ OstreCWEB/OstreCWEB/Views/StoryBuilder/CreatNewParagraph.cshtml
@@ -1,5 +1,5 @@
 ﻿@using OstreCWEB.DomainModels.StoryModels.Enums
-@model OstreCWEB.ViewModel.StoryBuilder.CreatNewParagraphView
+@model OstreCWEB.Services.StoryServices.ModelsView.CreatNewParagraphView
 
 @{
     ViewData["Title"] = "Create Paragraph";
