+++ OstreCWEB/OstreCWEB/Views/StoryBuilder/CreatParagraphFight.cshtml
@@ -1,4 +1,4 @@
-﻿@model OstreCWEB.ViewModel.StoryBuilder.CreatParagraphFightView
+﻿@model OstreCWEB.Services.StoryServices.ModelsView.CreatParagraphFightView
 
 @{
     ViewData["Title"] = "Create Paragraph Fight";
