+++ OstreCWEB/OstreCWEB/Views/StoryBuilder/CreatParagraphTest.cshtml
@@ -1,6 +1,6 @@
 ﻿@using OstreCWEB.DomainModels.CharacterModels.Enums
 @using OstreCWEB.DomainModels.StoryModels.Enums
-@model OstreCWEB.ViewModel.StoryBuilder.CreatParagraphTestView
+@model OstreCWEB.Services.StoryServices.ModelsView.CreatParagraphTestView
 
 @{
     ViewData["Title"] = "Create Paragraph Test";
