+++ OstreCWEB/OstreCWEB/Views/StoryBuilder/CreateStory.cshtml
@@ -1,4 +1,4 @@
-﻿@model OstreCWEB.ViewModel.StoryBuilder.StoriesView
+﻿@model OstreCWEB.Services.StoryServices.ModelsView.StoriesView
 
 @{
     ViewData["Title"] = "Create story";
