+++ OstreCWEB/OstreCWEB/Views/StoryBuilder/DeleteParagraph.cshtml
@@ -1,4 +1,4 @@
-﻿@model OstreCWEB.ViewModel.StoryBuilder.ParagraphElementView
+﻿@model OstreCWEB.Services.StoryServices.ModelsView.ParagraphElementView
 
 @{
     ViewData["Title"] = "Delete Paragraph";
