+++ OstreCWEB/OstreCWEB/Views/StoryBuilder/DeleteStory.cshtml
@@ -1,4 +1,4 @@
-﻿@model OstreCWEB.ViewModel.StoryBuilder.StoriesView
+﻿@model OstreCWEB.Services.StoryServices.ModelsView.StoriesView
 
 @{
     ViewData["Title"] = "Delete Story";
