+++ OstreCWEB/OstreCWEB/Views/StoryBuilder/EditChoice.cshtml
@@ -1,5 +1,5 @@
 ﻿@using OstreCWEB.DomainModels.CharacterModels.Enums
-@model OstreCWEB.ViewModel.StoryBuilder.ChoiceCreatorView
+@model OstreCWEB.Services.StoryServices.ModelsView.ChoiceCreatorView
 
 @{
     ViewData["Title"] = $"Edit Choice";
