+++ OstreCWEB/OstreCWEB/Views/StoryBuilder/EditParagraph.cshtml
@@ -1,6 +1,6 @@
 ﻿@using OstreCWEB.DomainModels.CharacterModels.Enums
 @using OstreCWEB.DomainModels.StoryModels.Enums
-@model OstreCWEB.ViewModel.StoryBuilder.EditParagraphView
+@model OstreCWEB.Services.StoryServices.ModelsView.EditParagraphView
 
 @{
     ViewData["Title"] = "Paragraph editor";
