+++ OstreCWEB/OstreCWEB/Views/StoryBuilder/EditStory.cshtml
@@ -1,4 +1,4 @@
-﻿@model OstreCWEB.ViewModel.StoryBuilder.StoriesView
+﻿@model OstreCWEB.Services.StoryServices.ModelsView.StoriesView
 
 @{
     ViewData["Title"] = "Edit Story";
