+++ OstreCWEB/OstreCWEB/Views/StoryBuilder/ParagraphDetails.cshtml
@@ -1,4 +1,4 @@
-﻿@model OstreCWEB.ViewModel.StoryBuilder.ParagraphDetailsView
+﻿@model OstreCWEB.Services.StoryServices.ModelsView.ParagraphDetailsView
 
 @{
     ViewData["Title"] = $"Story " + @Model.NameOfStory;
