+++ OstreCWEB/OstreCWEB/Views/StoryBuilder/StoryParagraphsList.cshtml
@@ -1,4 +1,4 @@
-﻿@model OstreCWEB.ViewModel.StoryBuilder.StoryParagraphsView 
+﻿@model OstreCWEB.Services.StoryServices.ModelsView.StoryParagraphsView
 @using OstreCWEB.DomainModels.StoryModels.Enums
 @{
     ViewData["Title"] = $"Story " + @Model.Name;
